Refactored state controllers for clarity
